refactor(canvas): normalize guide domain naming
