+++ packages/core/src/canvas/guides/geometry.ts
@@ -18,14 +18,6 @@ export interface GuideScreenSegment {
   y2: number
 }
 
-export interface GuideHit {
-  ownerId: string
-  guideId: string
-  axis: CanvasGuide['axis']
-  position: number
-  distance: number
-}
-
 export function getGuideScreenSegment(
   graph: SceneGraph,
   owner: SceneNode,
@@ -69,43 +61,4 @@ export function distanceToGuideSegment(x: number, y: number, segment: GuideScree
       ? 0
       : Math.max(0, Math.min(1, ((x - segment.x1) * dx + (y - segment.y1) * dy) / lengthSquared))
   return Math.hypot(x - (segment.x1 + t * dx), y - (segment.y1 + t * dy))
-}
-
-export function hitTestGuides(
-  graph: SceneGraph,
-  pageId: string,
-  viewport: GuideViewport,
-  x: number,
-  y: number,
-  tolerance = 5
-): GuideHit | null {
-  const page = graph.getNode(pageId)
-  if (!page) return null
-  let closest: GuideHit | null = null
-
-  const visit = (owner: SceneNode) => {
-    for (const guide of owner.guides) {
-      const distance = distanceToGuideSegment(
-        x,
-        y,
-        getGuideScreenSegment(graph, owner, guide, viewport)
-      )
-      if (distance <= tolerance && (!closest || distance < closest.distance)) {
-        closest = {
-          ownerId: owner.id,
-          guideId: guide.id,
-          axis: guide.axis,
-          position: guide.position,
-          distance
-        }
-      }
-    }
-    for (const childId of owner.childIds) {
-      const child = graph.getNode(childId)
-      if (child) visit(child)
-    }
-  }
-
-  visit(page)
-  return closest
 }
