+++ packages/core/src/canvas/index.ts
@@ -7,10 +7,9 @@ export {
 export {
   distanceToGuideSegment,
   getGuideScreenSegment,
-  hitTestGuides,
-  type GuideHit,
   type GuideScreenSegment,
   type GuideViewport
 } from './guides/geometry'
+export { hitTestGuides, type GuideHit } from './guides/hit-test'
 export type { GuideOverlayState, GuidePreview, GuideSelection } from './guides/types'
 export { SkiaRenderer, type RenderOverlays, type RulerTheme } from './renderer'
